+++ spec/config/tree_haver.rb
@@ -1,0 +1,3 @@
+# TreeHaver needs to be loaded early, so we can make the DependencyTags available
+require "tree_haver"
+require "tree_haver/rspec"